Refactored the lock file parsing and vulnerability downloading into separate modules
